% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLS.Score.R
\name{JLS.Score}
\alias{JLS.Score}
\title{Score Tests for Joint Location and Scale Models}
\usage{
JLS.Score(y, X = NULL, W = NULL, Z = NULL)
}
\arguments{
\item{y}{response vector}

\item{X}{vector (or matrix) of covariate(s) to be tested for "location" effect(s) on the response. If a matrix, each column represents a variable and each row represents an observation. This may be the same as W.}

\item{W}{vector (or matrix) of covariate(s) to be tested for "location" effect(s) on the response. If a matrix, each column represents a variable and each row represents an observation. This may be the same as X.}

\item{Z}{vector (or matrix) of covariate(s) to be included as adjustment variables in the "location" portion of the model. If a matrix, each column represents a variable and each row represents an observation.}
}
\value{
a table consisting of test statistics, degrees of freedom and p-vaules for the score test using the observed (W_obs) and expected (W_exp) information covariance matrix, and the robust covariance estimators V_A(D,I), V_B(D,E(I)), and V_C(E(D),E(I)).
}
\description{
This function performs the joint location and scale (JLS) or scale-only score tests (Soave, Lawless an Awadalla, submitted) to simultaneously test for mean and variance (or variance-only) associations with covariates.
}
\details{
No missing data are allowed - function will return an "error". Outcome must be quantitative and covariates may be discrete (categorical) or continuous.
}
\examples{
#################################################################################
## Example simulating data from Table 1, row 3 (Soave, Lawless, and Awadalla, submitted)
#################################################################################

#### Simulation parameters
n<-1000 # sample size
pX<-0.1 # covariate frequency

reps<-100000  # number of simulation replicates

#################################################################################
## Simulation replicates
#################################################################################

sims <-replicate(
 reps,
 expr = {
   XG<-rbinom(n,1,pX)
   y<-rnorm(length(XG),0,1)
   result_J<-JLS.Score(y=y,X=XG,W=XG,Z=NULL)
   result_S<-JLS.Score(y=y,X=NULL,W=XG,Z=XG)
   c(result_J[,3],result_S[,3])
 }
)

#################################################################################
## Power estimates
#################################################################################


alpha<-0.01
power<-rbind(apply(sims,1,function(x){sum(x<alpha)/length(x)}))
power.J<-power[,1:5]
power.S<-power[,6:10]
names(power.J)=names(power.S)=c("W_obs","W_exp","V_A","V_B","V_C")
power.J
power.S
}
\references{
Soave, D., Lawless, J.F., and Awadalla, P. Score tests for association in location and scale models. Submitted.
}
\author{
David Soave
}
\keyword{JLS}
\keyword{Score}
\keyword{test}
